+++ gazebo/gui/model/ModelEditorEvents.cc
@@ -42,8 +42,10 @@ event::EventT<void (std::string, std::string, std::string, std::string,
     std::string)> model::Events::jointChanged;
 event::EventT<void (std::string)> model::Events::openLinkInspector;
 event::EventT<void (std::string)> model::Events::openJointInspector;
+event::EventT<void (std::string)> model::Events::openModelPluginInspector;
 event::EventT<void (std::string, std::string)> model::Events::jointNameChanged;
 event::EventT<void (std::string)> model::Events::showLinkContextMenu;
 event::EventT<void (std::string)> model::Events::showJointContextMenu;
 event::EventT<void (std::string, bool)> model::Events::setSelectedLink;
 event::EventT<void (std::string, bool)> model::Events::setSelectedJoint;
+event::EventT<void (std::string)> model::Events::modelPluginInserted;